Modify the README.md and add some nonesense to it for no reason
So I just wanted to see if i can use "nano" to change anything in a .md file and i realised that i can. so that's happy news because I hate VScode.
I will change the readme file more and more throughout this project and reflect on this piece of shit course called "The Odin Project"